provide AVE to each file
Since SMODS is *hiding* load, we can't pass in an ENV. Therefore, we can't make AVE "global" to each loaded file. However, we can pass it in to the function call, then access the arguments provided through `...`. So each file has that appended to the top (although not all of them need it right now, but I'm future proofing; we might want it once I've had a chance to go through the other functions).